refactor cartelle e aggiunto VisualizzaBirreViewModel
