+++ app/src/main/res/layout/fragment_birra_dettagliata.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="2000dp"
-    tools:context=".ui.gestione_birra.BirraDettagliataFragment">
+    tools:context=".ui.visualizza_birre.BirraDettagliataFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
