+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -63,7 +63,7 @@
         tools:layout="@layout/fragment_cosa_devo_preparare_oggi" />
     <fragment
         android:id="@+id/birraDettagliataFragment"
-        android:name="it.unimib.brewday.ui.gestione_birra.BirraDettagliataFragment"
+        android:name="it.unimib.brewday.ui.visualizza_birre.BirraDettagliataFragment"
         android:label="fragment_birra_dettagliata"
         tools:layout="@layout/fragment_birra_dettagliata" >
         <argument
